+++ packages/ilmomasiina-backend/src/config.ts
@@ -140,6 +140,11 @@ const config = {
   /** Host for Mailgun API server. */
   mailgunHost: envString("MAILGUN_HOST", "api.eu.mailgun.net"),
 
+  /** How long each user has to edit their signup after creation. */
+  signupConfirmMins: envInteger("SIGNUP_CONFIRM_MINS", 30),
+  /** Whether signups can be edited for SIGNUP_CONFIRM_MINS after creation, even if signups for the event have closed. */
+  signupConfirmAfterClose: envBoolean("SIGNUP_CONFIRM_AFTER_CLOSE", false),
+
   /** How long after an event's date to remove signup details. */
   anonymizeAfterDays: envInteger("ANONYMIZE_AFTER_DAYS", 180),
   /** How long items stay in the database after deletion, in order to allow restoring accidentally deleted items. */