+++ packages/ilmomasiina-components/src/modules/editSignup/index.tsx
@@ -25,6 +25,7 @@ export function useEditSignupState({ id, editToken }: EditSignupProps) {
           [EDIT_TOKEN_HEADER]: editToken,
         },
       });
+      const now = Date.now();
       return {
         ...response,
         signup: {
@@ -33,21 +34,19 @@ export function useEditSignupState({ id, editToken }: EditSignupProps) {
           lastName: response.signup.lastName || "",
           email: response.signup.email || "",
         },
+        // Compute these once when the response arrives.
+        editingClosedOnLoad: response.signup.editableForMillis === 0,
+        confirmableUntil: now + response.signup.confirmableForMillis,
+        editableUntil: now + response.signup.editableForMillis,
       };
     },
     [id, editToken],
   );
 
-  // TODO: use data from server about editing end time
-  const registrationClosed =
-    !fetchSignup.result?.event.registrationEndDate ||
-    new Date(fetchSignup.result?.event.registrationEndDate) < new Date();
-
   return useShallowMemo<State>({
     editToken,
     pending: fetchSignup.pending,
     error: fetchSignup.error as ApiError | undefined,
-    registrationClosed,
     ...fetchSignup.result,
     isNew: fetchSignup.result && !fetchSignup.result.signup.confirmed,
   });